refactor(workbench): complete host package cutover
