--- apps/desktop/src/renderer/src/features/workspace-workbench/services/internal/workbenchCapabilityRegistry.ts
@@ -1,4 +1,0 @@
-export {
-  resolveWorkbenchCapabilityRegistry,
-  type WorkbenchCapabilityRegistryResult
-} from "@tutti-os/workbench-host";
