--- apps/desktop/src/renderer/src/features/workspace-workbench/services/internal/workbenchHostCoordinator.ts
@@ -1,8 +1,0 @@
-export {
-  createWorkbenchHostSessionConfiguration,
-  WorkbenchHostCoordinator,
-  type WorkbenchHostCoordinatorOptions,
-  type WorkbenchHostSessionConfiguration,
-  type WorkbenchHostSessionLease,
-  type WorkbenchHostSessionOpenInput
-} from "@tutti-os/workbench-host";
