+++ apps/desktop/src/renderer/src/features/workspace-workbench/services/internal/workbenchHostKernelBoundaries.test.ts
@@ -1,32 +1,30 @@
 import assert from "node:assert/strict";
-import { readFileSync } from "node:fs";
+import { existsSync, readFileSync } from "node:fs";
 import test from "node:test";
 
-const neutralKernelFiles = [
-  "workbenchCapabilityRegistry.ts",
-  "workbenchHostCoordinator.ts",
+const productHostContractFiles = [
   "workbenchHostPorts.ts",
-  "workbenchHostSession.ts",
   "workbenchProductProfile.ts"
 ] as const;
 
-const extractedKernelShimFiles = [
+const removedPrivateKernelFiles = [
   "workbenchCapabilityRegistry.ts",
   "workbenchHostCoordinator.ts",
   "workbenchHostSession.ts"
 ] as const;
 
-test("desktop private kernel paths delegate to the shared host package", () => {
-  for (const file of extractedKernelShimFiles) {
-    const source = readFileSync(new URL(`./${file}`, import.meta.url), "utf8");
-
-    assert.match(source, /from "@tutti-os\/workbench-host";/, file);
-    assert.doesNotMatch(source, /\b(?:class|function)\s+WorkbenchHost/, file);
+test("desktop has no private host-kernel compatibility paths", () => {
+  for (const file of removedPrivateKernelFiles) {
+    assert.equal(
+      existsSync(new URL(`./${file}`, import.meta.url)),
+      false,
+      file
+    );
   }
 });
 
-test("private workbench host kernel has no product, DI, or React runtime imports", () => {
-  for (const file of neutralKernelFiles) {
+test("desktop product host contracts have no DI or React runtime imports", () => {
+  for (const file of productHostContractFiles) {
     const source = readFileSync(new URL(`./${file}`, import.meta.url), "utf8");
 
     assert.doesNotMatch(
@@ -42,8 +40,8 @@ test("private workbench host kernel has no product, DI, or React runtime imports
   }
 });
 
-test("private workbench host kernel depends on surface contracts type-only", () => {
-  for (const file of neutralKernelFiles) {
+test("desktop product host contracts depend on surface contracts type-only", () => {
+  for (const file of productHostContractFiles) {
     const source = readFileSync(new URL(`./${file}`, import.meta.url), "utf8");
     const surfaceImports = source.match(
       /^import(?: type)?[^;]+from "@tutti-os\/workbench-surface";/gm
