--- apps/desktop/src/renderer/src/features/workspace-workbench/services/internal/workbenchHostSession.ts
@@ -1,9 +1,0 @@
-export {
-  areWorkbenchSnapshotPartitionsEqual,
-  WorkbenchHostSession,
-  type WorkbenchAuthenticatedPrincipalSnapshot,
-  type WorkbenchHostSessionOptions,
-  type WorkbenchHostSessionResolution,
-  type WorkbenchScope,
-  type WorkbenchSnapshotPartition
-} from "@tutti-os/workbench-host";
