+++ apps/desktop/src/renderer/src/features/workspace-workbench/services/workbenchHostCoordinator.interface.ts
@@ -1,5 +1,5 @@
 import { createDecorator } from "@tutti-os/infra/di";
-import type { WorkbenchHostCoordinator } from "./internal/workbenchHostCoordinator.ts";
+import type { WorkbenchHostCoordinator } from "@tutti-os/workbench-host";
 
 export const IWorkbenchHostCoordinator =
   createDecorator<WorkbenchHostCoordinator>("workbench-host-coordinator");
